+++ src/playground/positions/MidaPlaygroundPosition.ts
@@ -20,9 +20,7 @@
  * THE SOFTWARE.
 */
 
-import { MidaPlaygroundPositionParameters, } from "!/src/playground/positions/MidaPlaygroundPositionParameters";
 import { decimal, MidaDecimal, } from "#decimals/MidaDecimal";
-import { unsupportedOperationError, } from "#errors/MidaErrorUtilities";
 import { MidaOrder, } from "#orders/MidaOrder";
 import { MidaOrderDirection, } from "#orders/MidaOrderDirection";
 import { MidaPosition, } from "#positions/MidaPosition";
@@ -30,10 +28,21 @@ import { MidaPositionDirection, } from "#positions/MidaPositionDirection";
 import { MidaPositionStatus, } from "#positions/MidaPositionStatus";
 import { MidaProtection, } from "#protections/MidaProtection";
 import { MidaProtectionChange, } from "#protections/MidaProtectionChange";
+import { MidaProtectionChangeRejection, } from "#protections/MidaProtectionChangeRejection";
+import { MidaProtectionChangeStatus, } from "#protections/MidaProtectionChangeStatus";
+import { MidaProtectionDirectives, } from "#protections/MidaProtectionDirectives";
+import { MidaSymbol, } from "#symbols/MidaSymbol";
 import { MidaEmitter, } from "#utilities/emitters/MidaEmitter";
+import { MidaPlaygroundAccount, } from "!/src/playground/accounts/MidaPlaygroundAccount";
+import { MidaPlaygroundPositionParameters, } from "!/src/playground/positions/MidaPlaygroundPositionParameters";
+import { MidaPlaygroundTrade, } from "!/src/playground/trades/MidaPlaygroundTrade";
 
 export class MidaPlaygroundPosition extends MidaPosition {
     readonly #engineEmitter: MidaEmitter;
+    readonly #openingTrades: MidaPlaygroundTrade[];
+    readonly #closingTrades: MidaPlaygroundTrade[];
+    #realizedCommission: MidaDecimal;
+    #realizedProfit: MidaDecimal;
 
     public constructor ({
         id,
@@ -41,6 +50,7 @@ export class MidaPlaygroundPosition extends MidaPosition {
         tradingAccount,
         volume,
         direction,
+        entryPrice,
         protection,
         engineEmitter,
     }: MidaPlaygroundPositionParameters) {
@@ -49,20 +59,46 @@ export class MidaPlaygroundPosition extends MidaPosition {
             symbol,
             volume,
             direction,
+            entryPrice,
             tradingAccount,
             protection,
         });
 
         this.#engineEmitter = engineEmitter;
+        this.#openingTrades = [];
+        this.#closingTrades = [];
+        this.#realizedCommission = decimal(0);
+        this.#realizedProfit = decimal(0);
 
         this.#configureListeners();
     }
 
-    public override async getUsedMargin (): Promise<MidaDecimal> {
-        if (this.status === MidaPositionStatus.CLOSED) {
-            return decimal(0);
-        }
+    public override get tradingAccount (): MidaPlaygroundAccount {
+        return super.tradingAccount as MidaPlaygroundAccount;
+    }
 
+    public get openingTrades (): MidaPlaygroundTrade[] {
+        return this.#openingTrades;
+    }
+
+    public get closingTrades (): MidaPlaygroundTrade[] {
+        return this.#closingTrades;
+    }
+
+    public get trades (): MidaPlaygroundTrade[] {
+        return [ ...this.openingTrades, ...this.closingTrades, ];
+    }
+
+    public get realizedCommission (): MidaDecimal {
+        return this.#realizedCommission;
+    }
+
+    public get realizedProfit (): MidaDecimal {
+        return this.#realizedProfit;
+    }
+
+    // Leverage is not supported for now (please just set a high balance)
+    public override async getUsedMargin (): Promise<MidaDecimal> {
         return decimal(0);
     }
 
@@ -82,19 +118,14 @@ export class MidaPlaygroundPosition extends MidaPosition {
         });
     }
 
+    // TODO
     public override async getUnrealizedSwap (): Promise<MidaDecimal> {
-        if (this.status === MidaPositionStatus.CLOSED) {
-            return decimal(0);
-        }
-
         return decimal(0);
     }
 
+    // Commission is applied at every trade so total commission is always
+    // realized, use realizedCommission property
     public override async getUnrealizedCommission (): Promise<MidaDecimal> {
-        if (this.status === MidaPositionStatus.CLOSED) {
-            return decimal(0);
-        }
-
         return decimal(0);
     }
 
@@ -103,20 +134,123 @@ export class MidaPlaygroundPosition extends MidaPosition {
             return decimal(0);
         }
 
-        return decimal(0);
+        const entryPrice: MidaDecimal = this.entryPrice as MidaDecimal;
+        const completeSymbol: MidaSymbol = await this.tradingAccount.getSymbol(this.symbol) as MidaSymbol;
+        const [ bid, ask, ] = await Promise.all([ this.tradingAccount.getSymbolBid(this.symbol), this.tradingAccount.getSymbolAsk(this.symbol), ]);
+        let grossProfit: MidaDecimal;
+
+        if (this.direction === MidaPositionDirection.LONG) {
+            grossProfit = bid.sub(entryPrice);
+        }
+        else {
+            grossProfit = entryPrice.sub(ask);
+        }
+
+        return grossProfit.mul(this.volume).mul(completeSymbol.lotUnits);
     }
 
-    public override async changeProtection (protection: MidaProtection): Promise<MidaProtectionChange> {
-        throw unsupportedOperationError(this.tradingAccount.platform);
+    public override async changeProtection (protection: MidaProtectionDirectives): Promise<MidaProtectionChange> {
+        const requestedProtection: MidaProtectionDirectives = protection;
+        const [ bid, ask, ] =
+                await Promise.all([ this.tradingAccount.getSymbolBid(this.symbol), this.tradingAccount.getSymbolAsk(this.symbol), ]);
+        let rejected: boolean = false;
+
+        // <protection-validation>
+        if (this.direction === MidaPositionDirection.LONG) {
+            if ("stopLoss" in requestedProtection && decimal(requestedProtection.stopLoss).greaterThanOrEqual(bid)) {
+                rejected = true;
+            }
+
+            if ("takeProfit" in requestedProtection && decimal(requestedProtection.takeProfit).lessThanOrEqual(bid)) {
+                rejected = true;
+            }
+        }
+        else {
+            if ("stopLoss" in requestedProtection && decimal(requestedProtection.stopLoss).lessThanOrEqual(ask)) {
+                rejected = true;
+            }
+
+            if ("takeProfit" in requestedProtection && decimal(requestedProtection.takeProfit).greaterThanOrEqual(ask)) {
+                rejected = true;
+            }
+        }
+
+        if (rejected) {
+            return {
+                status: MidaProtectionChangeStatus.REJECTED,
+                rejection: MidaProtectionChangeRejection.INVALID_PROTECTION,
+                requestedProtection,
+            };
+        }
+        // </protection-validation>
+
+        await this.onProtectionChange(requestedProtection as MidaProtection);
+
+        return {
+            status: MidaProtectionChangeStatus.SUCCEEDED,
+            requestedProtection,
+        };
     }
 
     #configureListeners (): void {
-        this.#engineEmitter.on("trade", (event) => {
+        this.#engineEmitter.on("trade", async (event) => {
             const { trade, } = event.descriptor;
 
-            if (trade.positionId === this.id) {
-                this.onTrade(trade);
+            if (trade.positionId !== this.id) {
+                return;
             }
+
+            this.#realizedCommission =
+                    this.#realizedCommission.add(trade.commission);
+
+            if (trade.isOpening) {
+                this.#openingTrades.push(trade);
+            }
+            else {
+                const profit = await this.getUnrealizedGrossProfit();
+                this.#realizedProfit =
+                        this.#realizedProfit.add(profit.mul(trade.volume.div(this.volume)));
+
+                this.#closingTrades.push(trade);
+            }
+
+            this.onTrade(trade);
+            this.#updateEntryPrice();
         });
+    }
+
+    #updateEntryPrice (): void {
+        let pivotVolume: MidaDecimal = this.volume;
+
+        if (this.status === MidaPositionStatus.CLOSED) {
+            this.onEntryPriceUpdate(undefined);
+
+            return;
+        }
+
+        const openingTrades: MidaPlaygroundTrade[] = this.openingTrades;
+        let entryPrice: MidaDecimal = decimal(0);
+
+        // @ts-ignore
+        openingTrades.sort((a: MidaPlaygroundTrade, b: MidaPlaygroundTrade) => a.executionDate.timestamp - b.executionDate.timestamp);
+
+        for (const trade of openingTrades) {
+            const remainingVolume: MidaDecimal = pivotVolume.sub(trade.volume);
+
+            if (remainingVolume.lessThan(0)) {
+                entryPrice = entryPrice.add(pivotVolume.mul(trade.executionPrice as MidaDecimal));
+                pivotVolume = decimal(0);
+            }
+            else {
+                entryPrice = entryPrice.add(trade.volume.mul(trade.executionPrice as MidaDecimal));
+                pivotVolume = remainingVolume;
+            }
+        }
+
+        if (!pivotVolume.isZero()) {
+            console.log("Error during position entry price calculation");
+        }
+
+        this.onEntryPriceUpdate(entryPrice.div(this.volume));
     }
 }