+++ tests/playground/MidaPlaygroundAccount.spec.ts
@@ -20,17 +20,20 @@
  * THE SOFTWARE.
 */
 
-import { MidaPlaygroundEngine, } from "!/src/playground/MidaPlaygroundEngine";
 import { decimal, MidaDecimal, } from "#decimals/MidaDecimal";
 import { MidaOrderDirection, } from "#orders/MidaOrderDirection";
 import { MidaOrderRejection, } from "#orders/MidaOrderRejection";
 import { MidaOrderStatus, } from "#orders/MidaOrderStatus";
+import { MidaPosition, } from "#positions/MidaPosition";
+import { MidaPositionDirection, } from "#positions/MidaPositionDirection";
+import { MidaPositionStatus, } from "#positions/MidaPositionStatus";
 import { MidaTick, } from "#ticks/MidaTick";
 import { readTicksFromFile, } from "#utilities/MidaFileSystem";
+import { MidaPlaygroundEngine, } from "!/src/playground/MidaPlaygroundEngine";
+import { MidaPlaygroundPosition, } from "!/src/playground/positions/MidaPlaygroundPosition";
 
 // eslint-disable-next-line max-lines-per-function
 describe("MidaPlaygroundAccount", () => {
-    const ticks: MidaTick[] = [];
     const symbolParameters = {
         symbol: "ETHUSD",
         baseAsset: "ETH",
@@ -40,28 +43,24 @@ describe("MidaPlaygroundAccount", () => {
         minLots: decimal(0),
         maxLots: decimal(0),
         lotUnits: decimal(0),
-        pipPosition: -1,
+        pipPosition: 2,
         digits: 2,
     };
 
-    beforeAll(async () => {
-        // First tick date is 2022-01-01T14:00:00.000Z
-        // Last tick date is 2022-01-01T14:01:57.143Z
-        for await (const tick of readTicksFromFile("./series/ETHUSD.csv", "ETHUSD")) {
-            if (tick) {
-                ticks.push(tick);
-            }
-        }
-    });
+    async function createEngine (): Promise<MidaPlaygroundEngine> {
+        const engine = new MidaPlaygroundEngine({ localDate: "2022-01-01T00:00:00.000Z", });
+        const ticksGenerator: AsyncGenerator<MidaTick | undefined> = readTicksFromFile("./series/ETHUSD.csv", "ETHUSD");
+
+        engine.setTicksGenerator("ETHUSD", ticksGenerator);
+        await engine.elapseTicks(1);
+
+        return engine;
+    }
 
     // eslint-disable-next-line max-lines-per-function
     describe(".placeOrder", () => {
-        it("P/L is equal to spread when buying and selling at the same tick", async () => {
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
-
-            engine.addSymbolTicks("ETHUSD", ticks);
-            await engine.elapseTicks(1);
-
+        it("p/l is equal to spread when buying and selling at the same tick", async () => {
+            const engine = await createEngine();
             const currentAsk: MidaDecimal = await engine.getSymbolAsk("ETHUSD");
             const account = await engine.createAccount({
                 balanceSheet: {
@@ -90,20 +89,17 @@ describe("MidaPlaygroundAccount", () => {
             expect((await account.getBalance()).equals(expectedBalance)).toBe(true);
         });
 
-        it("P/L is equal to spread + commission when buying and selling at the same tick", async () => {
+        it("p/l is equal to spread + commission when buying and selling at the same tick", async () => {
+            const engine = await createEngine();
             const fixedCommission: MidaDecimal = decimal(10);
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
-
-            engine.setCommissionCustomizer(async () => [ "USD", fixedCommission, ]);
-            engine.addSymbolTicks("ETHUSD", ticks);
-            await engine.elapseTicks(1);
-
             const currentAsk: MidaDecimal = await engine.getSymbolAsk("ETHUSD");
             const account = await engine.createAccount({
                 balanceSheet: {
                     "USD": currentAsk,
                 },
             });
+
+            engine.setCommissionCustomizer(async () => [ "USD", fixedCommission, ]);
 
             await account.addSymbol(symbolParameters);
             await account.placeOrder({
@@ -123,15 +119,45 @@ describe("MidaPlaygroundAccount", () => {
             const expectedBalance: MidaDecimal = currentAsk.subtract(currentAsk.subtract(await engine.getSymbolBid("ETHUSD")));
 
             // eslint-disable-next-line max-len
-            expect((await account.getBalance()).equals(expectedBalance.subtract(fixedCommission.multiply(2)))).toBe(true);
+            expect((await account.getBalance()).equals(expectedBalance.subtract(fixedCommission.mul(2)))).toBe(true);
         });
 
-        it("BUY LIMIT is executed at the expected price", async () => {
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
+        it("position p/l is equal to spread when buying and selling at the same tick", async () => {
+            const engine = await createEngine();
+            const bid = await engine.getSymbolBid("ETHUSD");
+            const ask = await engine.getSymbolAsk("ETHUSD");
+            const spread = ask.sub(bid);
+            const account = await engine.createAccount({
+                balanceSheet: {
+                    "USD": ask,
+                },
+            });
 
-            engine.addSymbolTicks("ETHUSD", ticks);
-            await engine.elapseTicks(1);
+            await account.addSymbol(symbolParameters);
 
+            const openOrder = await account.placeOrder({
+                symbol: "ETHUSD",
+                direction: MidaOrderDirection.BUY,
+                volume: 1,
+            });
+
+            const position = await openOrder.getPosition() as MidaPlaygroundPosition;
+
+            expect((await position.getUnrealizedGrossProfit()).eq(spread.mul(-1)));
+            expect(position.realizedProfit.eq(0));
+
+            await account.placeOrder({
+                symbol: "ETHUSD",
+                direction: MidaOrderDirection.SELL,
+                volume: 1,
+            });
+
+            expect((await position.getUnrealizedGrossProfit()).eq(0));
+            expect(position.realizedProfit.eq(spread.mul(-1)));
+        });
+
+        it("buy limit is executed at the expected price", async () => {
+            const engine = await createEngine();
             const account = await engine.createAccount({
                 balanceSheet: {
                     "USD": 100000,
@@ -162,12 +188,8 @@ describe("MidaPlaygroundAccount", () => {
             expect(limitOrder.executionPrice?.equals(limitPrice)).toBe(true);
         });
 
-        it("BUY STOP is executed at the expected price", async () => {
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
-
-            engine.addSymbolTicks("ETHUSD", ticks);
-            await engine.elapseTicks(1);
-
+        it("buy stop is executed at the expected price", async () => {
+            const engine = await createEngine();
             const account = await engine.createAccount({
                 balanceSheet: {
                     "USD": 100000,
@@ -197,15 +219,54 @@ describe("MidaPlaygroundAccount", () => {
             expect(limitOrder.executionPrice?.equals(stopPrice)).toBe(true);
         });
 
-        it("BUY is rejected when owned quote currency volume is not enough", async () => {
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
+        it("position is closed with take profit at the expected price #1", async () => {
+            const engine = await createEngine();
+            const account = await engine.createAccount({
+                balanceSheet: {
+                    "USD": 100000,
+                },
+            });
 
-            engine.addSymbolTicks("ETHUSD", ticks);
+            await account.addSymbol(symbolParameters);
+
+            const takeProfit = 3707.6;
+            const order = await account.placeOrder({
+                symbol: "ETHUSD",
+                direction: MidaOrderDirection.BUY,
+                volume: 1,
+                protection: {
+                    takeProfit, // Tick #41
+                },
+            });
+
+            expect(order.status).toBe(MidaOrderStatus.EXECUTED);
+
+            const position: MidaPosition = await order.getPosition() as MidaPosition;
+
+            expect(position.status).toBe(MidaPositionStatus.OPEN);
+            expect(position.direction).toBe(MidaPositionDirection.LONG);
+            expect(position.entryPrice?.eq(order.executionPrice as MidaDecimal)).toBe(true);
+            expect(position.takeProfit?.eq(takeProfit)).toBe(true);
+
+            await engine.elapseTicks(10);
+            await engine.elapseTicks(10);
+            await engine.elapseTicks(19);
+
+            expect(position.status).toBe(MidaPositionStatus.OPEN);
+            expect(position.direction).toBe(MidaPositionDirection.LONG);
+
             await engine.elapseTicks(1);
 
+            expect(position.status).toBe(MidaPositionStatus.CLOSED);
+            expect(position.direction).toBe(undefined);
+        });
+
+        it("buy is rejected when owned quote currency volume is not enough", async () => {
+            const engine = await createEngine();
             const account = await engine.createAccount();
 
             await account.addSymbol(symbolParameters);
+
             const order = await account.placeOrder({
                 symbol: "ETHUSD",
                 direction: MidaOrderDirection.BUY,
@@ -216,12 +277,8 @@ describe("MidaPlaygroundAccount", () => {
             expect(order.rejection).toBe(MidaOrderRejection.NOT_ENOUGH_MONEY);
         });
 
-        it("SELL is rejected when owned base currency volume is not enough", async () => {
-            const engine = new MidaPlaygroundEngine({ localDate: "2022-04-04T04:04:04.004Z", });
-
-            engine.addSymbolTicks("ETHUSD", ticks);
-            await engine.elapseTicks(1);
-
+        it("sell is rejected when owned base currency volume is not enough", async () => {
+            const engine = await createEngine();
             const account = await engine.createAccount();
 
             await account.addSymbol(symbolParameters);